moved helpers to a lib folder.
created suite for builtin colors
