+++ src/test/built-in-colors.test.ts
@@ -1,9 +1,35 @@
 import * as vscode from 'vscode';
 import * as assert from 'assert';
-import { Commands, ColorSettings, BuiltInColors } from '../models';
-import { getPeacockWorkspaceConfig } from './helpers';
+import {
+  Commands,
+  ColorSettings,
+  BuiltInColors,
+  IPeacockSettings
+} from '../models';
+import { getPeacockWorkspaceConfig } from './lib/helpers';
+import {
+  teardownTestSuite,
+  setupTestSuite
+} from './lib/setup-teardown-test-suite';
 
-export function testChangingColorToAngularRed():
+suite('can set color to built-in color', () => {
+  let extension: vscode.Extension<any>;
+  let originalValues = <IPeacockSettings>{};
+
+  suiteSetup(async () => {
+    extension = await setupTestSuite(extension, originalValues);
+  });
+
+  test('can set color to Angular Red', testChangingColorToAngularRed());
+
+  test('can set color to Vue Green', testChangingColorToVueGreen());
+
+  test('can set color to React Blue', testChangingColorToReactBlue());
+
+  suiteTeardown(() => teardownTestSuite(originalValues));
+});
+
+function testChangingColorToAngularRed():
   | ((this: Mocha.ITestCallbackContext, done: MochaDone) => any)
   | undefined {
   return testBuiltInColor(
@@ -11,12 +37,12 @@ export function testChangingColorToAngularRed():
     BuiltInColors.Angular
   );
 }
-export function testChangingColorToVueGreen():
+function testChangingColorToVueGreen():
   | ((this: Mocha.ITestCallbackContext, done: MochaDone) => any)
   | undefined {
   return testBuiltInColor(Commands.changeColorToVueGreen, BuiltInColors.Vue);
 }
-export function testChangingColorToReactBlue():
+function testChangingColorToReactBlue():
   | ((this: Mocha.ITestCallbackContext, done: MochaDone) => any)
   | undefined {
   return testBuiltInColor(Commands.changeColorToReactBlue, BuiltInColors.React);
